decouple ui from game logic
